Update README.md
As A Recent Graduate With A Solid Foundation In Python Programming, I Am Actively Seeking A Position As A Python Developer. Armed With A [Your Degree And University/College], I Have Acquired Hands-On Experience Through Academic Projects, Demonstrating Proficiency In [Mention Any Specific Skills Or Frameworks You've Learned]. Eager To Contribute My Enthusiasm And Fresh Perspective To A Dynamic Development Team, I Am Committed To Continuous Learning And Excited About The Opportunity To Apply My Skills To Real-World Projects As I Embark On A Fulfilling Career As A Python Developer.